+++ src/org/red5/io/object/UnsignedShort.java
@@ -75,12 +75,12 @@ public final class UnsignedShort extends UnsignedNumber {
 
 	@Override
 	public double doubleValue() {
-		return (double) value;
+		return value;
 	}
 
 	@Override
 	public float floatValue() {
-		return (float) value;
+		return value;
 	}
 
 	@Override
@@ -125,7 +125,7 @@ public final class UnsignedShort extends UnsignedNumber {
 	
 	@Override
 	public int hashCode() {
-		return (int)value;
+		return value;
 	}
 
 	@Override
